requirement.txt
python==3.8
ffmpeg: environment required package dependency
setuptools-rust==4.7.1
openai-whisper



command:
for Text2Voice
brew install mpv # we have window mpv and mac mpv
edge-tts
brew

Audio
sudo apt install ffmpeg
pip install torchaudio ipywebrtc notebook
jupyter nbextension enable --py widgetsnbextension



For speech recognition


hugging face, suno, speach5 tts, whisper, whisper-medium

